+++ CHANGELOG.md
@@ -10,16 +10,22 @@
  - App does not crash at first connection to laundry
  - Var. IOS bugs
  - Timetable was only shown when rules were set
+ - Disable back arrow in timetable
+ - Make login scrollable
+ - Scroll table when modal present
 
 ### Added
-
+ - Images
  - Date-picker for timetable
  - Modal for confirmation
  - Scroll to active time
  - Show current time
+ - Add icons to timetable
+ - Add error message to failed login
 
 ### Changed
 
+ - Redesign login page
  - Redesign machine name
  - Redesign settings page